Use config to control magics registry
